reubico post de propiedades
